Rename inner class
